Realign with Unexpected.
This commit ontains two careful semver major changes which adjust
the exposed APIs such that we claw back some flexibility to allow
moving forward.

In particular:
* Reserve subsequent arguments to the expect() function
  While expect.js did not use these, its de facto heir developed
  within the chai project does. If we are to grow support for it
  we must be in a position to repurpose those arguments.
* Redefine addAssertion() so it is Unexpected compatible
  The library defined its own addAssertion() API which was not
  compatible with Unexpected core. This meant an inability to
  expose an interface that was compatible and thus present no
  obstacles to the use of Unexpected plugins with the library.

Of particular note is that the second of the two changes actually
introduces the ability to automatically register an Unexpected
assertion as an expect.js assertion.

The changes have been done made with particular care for existing
users who will have clear messages explaining the situation.