Rewrite the way steps are defined.
Rather than try to define paths, move to an internal architecture that
treats the properties in an assertion chains as a bag of flags which
will apply when an assertion function is invokved.

Use this opportunty to stop creating new objects every time a property
is dereferenced. Instead, record flags being set on the current chain
and only clean this when chaining.

This commit also splits the target on which methods will be attached
from the "scope" which maintians the flags. Move to always attaching
getters for the assertions upfront which ensures error wrapping and
value chaining will function correctly at every level.