Cleave ExpectFacade and Assertion files and expose a chai entrypoint.
Rearrange things to consolidate assertion implementations into a
single file. Use the opportunity to also split the creation of the
wrapped properties to isolate concerns.

At the top level, expose a chai compatible object that exposes the
expect property and the definitions of the assertions (the least
amount required to potentially support plugins).

As of this commit the library is a drop-in replacement for chai.